Rename generate-div-chunks.js to generate-bg-rows.js, etc.
I renamed generate-div-chunks.js to generate-bg-rows.js because I think that makes it more obvious what its purpose is - it generates the rows of divs in the site that offset the rainbow background.
I also renamed a few variables and edited some comments to reflect this.
Additionally, I changed the line `for (rowNum = 1; rowNum < bgRowHeight; rowNum++) {` to `for (rowNum = 1; rowNum < 21; rowNum++) {`. I had gotten the 21 in `bgRowHeight` mixed up with the 20 iterations it takes to loop around the class numbers in the CSS. In case you were wondering, the reason it was 21 and not 20 was because the index of the loop started at 1 because the CSS did.