Switch from raster to vector for text!
It took a lot of work (retrofitting), but I've finally accomplished it: h now uses a procedurally-generated SVG to show the text. It no longer needs a hardcoded PNG file. This means CUSTOM TEXT IS BACK-ish. I still need to figure out a user-friendly way to enter that custom text into the site...
But this is a big milestone. This joke website it shaping up to be even better.
A huge thank you to the person/people behind https://finalboss.org. They are my inspiration.