Make favicon look better
Early on, I made the favicon the h.png file that the original website used for the actual h's. It wasn't intended to be a favicon, and it wasn't the right aspect ratio. Also, the h was transparent, so that caused problems when the background behind it was really anything other than white.
This new one uses the SVG from the new website, modified to be bigger and in 1:1 aspect ratio.